🚚 [projects] Move function `linkproject` from `services` to `link` module
